added images & proper font
