+++ src/components/layout.module.css
@@ -74,3 +74,7 @@
   margin-bottom: 0px;
   color: #da300f;
 }
+
+div {
+  font-family: "Poppins Regular";
+}
