+++ src/pages/callback.js
@@ -1,4 +1,3 @@
-// src/pages/callback.js
 import React, { useEffect, useState } from "react"
 import { navigate } from "gatsby"
 
@@ -68,7 +67,7 @@ const CallbackPage = () => {
 
         if (!response.ok) {
           setStatus(
-            `Token exchange failed: ${response.status} - ${responseText}`,
+            `Token exchange failed: ${response.status} - ${responseText}`
           )
           return
         }
@@ -89,7 +88,7 @@ const CallbackPage = () => {
           localStorage.setItem("trakt_refresh_token", data.refresh_token)
           localStorage.setItem(
             "trakt_token_expiry",
-            new Date(Date.now() + data.expires_in * 1000).toISOString(),
+            new Date(Date.now() + data.expires_in * 1000).toISOString()
           )
 
           setStatus("Successfully authenticated!")
