+++ src/pages/now.js
@@ -1,5 +1,20 @@
 import React, { useEffect, useState } from "react"
 import PropTypes from "prop-types"
+import Layout from "../components/layout"
+import {
+  BookDisplay,
+  WatchDisplay,
+  TrackDisplay,
+} from "../components/mediaDisplay"
+
+export const Head = () => {
+  return (
+    <>
+      <title>{"Brendan Reed"}</title>
+      <html lang="en" />
+    </>
+  )
+}
 
 // Media Section Components
 const MediaSection = ({ title, error, loading, children }) => (
@@ -22,121 +37,6 @@ MediaSection.propTypes = {
   error: PropTypes.string,
   loading: PropTypes.bool,
   children: PropTypes.node,
-}
-
-// Book Components
-const BookDisplay = ({ book, type }) => (
-  <div className="mb-3 rounded border border-gray-600 p-3">
-    <div className="font-medium text-gray-300">{book.title}</div>
-    <div className="text-sm text-gray-400">by {book.author}</div>
-    {type === "read" && book.dateRead && (
-      <div className="mt-1 text-xs text-gray-500">
-        Finished: {new Date(book.dateRead).toLocaleDateString()}
-      </div>
-    )}
-  </div>
-)
-
-BookDisplay.propTypes = {
-  book: PropTypes.shape({
-    title: PropTypes.string.isRequired,
-    author: PropTypes.string.isRequired,
-    dateRead: PropTypes.string,
-  }).isRequired,
-  type: PropTypes.oneOf(["current", "read"]).isRequired,
-}
-
-// Watch History Components
-const WatchDisplay = ({ item, type }) => (
-  <div className="mb-3 rounded border border-gray-600 p-3">
-    {type === "tv" ? (
-      <>
-        <div className="font-medium text-gray-300">{item.show.title}</div>
-        <div className="text-sm text-gray-400">
-          S{item.episode.season}E{item.episode.number} - {item.episode.title}
-        </div>
-      </>
-    ) : (
-      <div className="font-medium text-gray-300">
-        {item.movie.title} ({item.movie.year})
-      </div>
-    )}
-    <div className="mt-1 text-xs text-gray-500">
-      Watched: {new Date(item.watched_at).toLocaleDateString()}
-    </div>
-  </div>
-)
-
-WatchDisplay.propTypes = {
-  item: PropTypes.shape({
-    show: PropTypes.shape({
-      title: PropTypes.string,
-    }),
-    episode: PropTypes.shape({
-      season: PropTypes.number,
-      number: PropTypes.number,
-      title: PropTypes.string,
-    }),
-    movie: PropTypes.shape({
-      title: PropTypes.string,
-      year: PropTypes.number,
-    }),
-    watched_at: PropTypes.string.isRequired,
-  }).isRequired,
-  type: PropTypes.oneOf(["tv", "movies"]).isRequired,
-}
-
-// Music Components
-const TrackDisplay = ({ track, type }) => (
-  <div className="mb-3 rounded border border-gray-600 p-3">
-    <div className="flex items-start justify-between">
-      <div>
-        <div className="font-medium text-gray-300">{track.name}</div>
-        <div className="text-sm text-gray-400">by {track.artist}</div>
-        {track.album && (
-          <div className="text-sm text-gray-500">{track.album}</div>
-        )}
-      </div>
-      <div className="text-right">
-        {type === "recent" &&
-          (track.isNowPlaying ? (
-            <span className="rounded-full bg-emerald-600 px-2 py-1 text-xs text-white">
-              Now Playing
-            </span>
-          ) : (
-            <span className="text-sm text-gray-500">
-              {track.timestamp ? formatTimestamp(track.timestamp) : ""}
-            </span>
-          ))}
-        {type === "top" && (
-          <span className="text-sm text-gray-500">{track.playcount} plays</span>
-        )}
-      </div>
-    </div>
-  </div>
-)
-
-TrackDisplay.propTypes = {
-  track: PropTypes.shape({
-    name: PropTypes.string.isRequired,
-    artist: PropTypes.string.isRequired,
-    album: PropTypes.string,
-    isNowPlaying: PropTypes.bool,
-    timestamp: PropTypes.number,
-    playcount: PropTypes.number,
-  }).isRequired,
-  type: PropTypes.oneOf(["recent", "top"]).isRequired,
-}
-
-const formatTimestamp = timestamp => {
-  const date = new Date(timestamp)
-  const now = new Date()
-  const diff = now - date
-
-  if (diff < 60000) return "Just now"
-  if (diff < 3600000) return `${Math.floor(diff / 60000)}m ago`
-  if (diff < 86400000) return `${Math.floor(diff / 3600000)}h ago`
-  return date.toLocaleDateString()
 }
 
 // Main Page Component
@@ -168,7 +68,7 @@ const NowPage = () => {
         .finally(() => setWatchLoading(false))
     } else {
       setWatchError(
-        "No authentication token found. Please connect your Trakt account.",
+        "No authentication token found. Please connect your Trakt account."
       )
       setWatchLoading(false)
     }
@@ -189,7 +89,7 @@ const NowPage = () => {
   }, [])
 
   return (
-    <div>
+    <Layout>
       <div className="min-h-screen p-6 font-sans text-white">
         <h1 className="mb-6 text-4xl font-normal">{"What I'm Up To Now"}</h1>
 
@@ -273,7 +173,7 @@ const NowPage = () => {
           </div>
         </div>
       </div>
-    </div>
+    </Layout>
   )
 }
